// src/common/enums/course-message.enum.ts
export enum CourseMessages {
    NOT_FOUND = 'Course not found',
    ALREADY_EXISTS = 'Course already exists',
    CREATED = 'Course created successfully',
    UPDATED = 'Course updated successfully',
    DELETED = 'Course deleted successfully',

    UNAUTHORIZED_ACCESS = 'Unauthorized access',
    INVALID_CREDENTIALS = 'Invalid credentials',
}